remove functions from main
